+++ src/entities/transformer.ts
@@ -23,10 +23,41 @@ export default async function transform(code: string) {
 							path.replaceWith(babel.types.awaitExpression(path.node));
 						}
 					},
+					// Endless loop protection
+					WhileStatement(path) {
+						if (path.node.body.type === "BlockStatement") {
+							path.node.body.body.unshift(
+								babel.types.expressionStatement(
+									babel.types.awaitExpression(
+										babel.types.newExpression(
+											babel.types.identifier("Promise"),
+											[
+												babel.types.identifier("setTimeout"),
+											]
+										)
+									)
+								)
+							);
+						}
+					},
+					ForStatement(path) {
+						if (path.node.body.type === "BlockStatement") {
+							path.node.body.body.unshift(
+								babel.types.expressionStatement(
+									babel.types.awaitExpression(
+										babel.types.newExpression(
+											babel.types.identifier("Promise"),
+											[
+												babel.types.identifier("setTimeout"),
+											]
+										)
+									)
+								)
+							);
+						}
+					}
 				},
 			},
 		],
-		sourceMaps: true,
-		filename: "script.js",
 	}))!;
 }